+++ source/photon.h
@@ -16,7 +16,7 @@ void photon_free();
 int photon_get_nb();
 void photon_update();
 int photon_retirer(int id);
-void photon_destruction_dehors(int xmin, int xmax,int ymin, int ymax);
+void photon_destruction_dehors(double xmin, double xmax,double ymin, double ymax);
 void photon_trajectoire_reflechie(PHOTON* ph, SEGMENT s_reflecteur);
 int photon_add_ph(PHOTON ph);
 void photon_add(VECTEUR position, double angle);
